refactor : repository id type 수정 - TeacherRepository : Long -> Integer - StudentRepository : Long -> Integer - 디렉토리 rename : domain -> entity

+++ src/main/java/com/monari/monariback/student/repository/StudentRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.monari.monariback.student.domain.Student;
+import com.monari.monariback.student.entity.Student;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
 	@Query("""
 			select s from Student s
 			WHERE s.socialId = :socialId
